upgrade Fabric8 Maven plugin to the 3.2 series
The Fabric8 Maven plugin 3.1 adds an ImageStream definition
to target/classes/META-INF/fabric8/openshift.yml, which means
that the OpenShiftTestAssistant deletes it after the test.
This then causes failures in a follow-up fabric8:deploy.

The Fabric8 Maven plugin 3.2 doesn't do that, which is good.
However, it tries to apply the "webapp" generator, which fails
the build because this generator doesn't support S2I. Fortunately,
this generator isn't needed, so it can be just excluded.

Also contains other small improvements of the POM.